Fix four review defects, two CI findings, and the P4 boot loop
Review and CI found six real bugs in the three commits before this one, four of
them invisible to 1329 passing tests. The engine also shrank from 1440 to 784
bytes, which is what was boot-looping the P4.

Perf: no per-tick cost. The engine is 656 bytes smaller PER SCRIPTED MODULE
(three per pipeline), which is stack at boot and heap for the life of the device.

Core
- A member assignment stored the WRONG REGISTER once the spill pass engaged.
  StoreCtrl/StoreCtrl16 reported kArg4 as their first source, and the rewriter
  writes sources back positionally, so the value moved to `b` while `a` became
  kArg4's register; both lowerings read the value from `a`. Shipped Xtensa
  scripts already reach the spill threshold. This is the same trap this project
  documented in its own backlog after hitting it on LoadIdx/StoreIdx, left in the
  sibling case: the fix there was to report only the real vreg operands, and it
  is the fix here.
- The seeding mask was a uint32_t while member offsets run 0..63. A shift of 32
  or more is undefined behaviour and in practice aliased mod 32, so a member
  above offset 31 was never recorded as seeded: it snapped back to its
  initializer on every recompile, losing the live value the mechanism exists to
  keep, and corrupted the bit of the member it aliased. The mask was written when
  the budget was 16 bytes and did not grow with it; a static_assert now ties the
  two together.
- An ARRAY was reseeded only at element 0, so after an edit that moved it,
  elements 1..n held the previous program's bytes. Every element is seeded now,
  which is what "an array starts at zero" has to mean.
- sizeof(MoonLive) was 1440 bytes, held BY VALUE in every scripted module and
  constructed on the main task's stack by registerType's probe. 876 of that was a
  seeded-name table indexed by ARENA BYTE (64 rows) when a class can hold at most
  8 members. Indexed by member instead: 784 bytes. THIS WAS THE P4 BOOT LOOP; the
  board now boots and runs at ~140 fps.
- A uint16_t member's initializer was cast to a byte on the way in, so
  `uint16_t phase = 1000;` started at 232. Invisible to every existing test,
  which observe through setRGB's byte truncation while the error is always a
  multiple of 256.
- ControlType::FilePath's picker is a FIXED-SIZE type. A caller passing two
  elements where three are read compiled fine and read past the end (ASan caught
  it). The parameter refuses it now, which is where a fixed-size contract belongs.
- A name-length loop read `name[n]` before testing the bound, so a name filling
  its buffer without a terminator was read one past the end (CodeQL, high).
- A file DELETE now fires the re-derive seam, as a write does. The seam's own
  rationale says the rule belongs in core rather than in whichever client
  remembers a follow-up, and a deleted file is a change to persistent state.

Light domain
- Dead code removed: scriptHashChunk (its only caller was scriptHash, and the
  chunked walk it existed for is gone) and MoonLive::stringBytes (no callers).
- Comments that contradicted their own implementation: three claimed
  scriptFileHash reads "through a small stack buffer and allocates nothing" while
  the code below did one whole-file read. Also the layout's onControlChanged doc
  described clearing a hash the body no longer clears, the modifier's described a
  four-argument setXYZ that no longer exists, and the effect carried a
  default-script comment for a default that was removed.

UI
- The dirty dot and the disabled style did NOTHING on a card: the CSS targets
  .fm-tool and the card mounts a .card-btn. Both are covered now.
- A FAILED save was silent. The card has no status line by design (the module's
  own status is authoritative and the device writes it), so a failed write now
  alerts rather than leaving the dot as the only signal.
- currentColor -> currentcolor (Stylelint).

Docs
- The P4's remaining crash is backlogged with its evidence: it panics with
  `Cache error` every few minutes while IDLE, and it is pre-existing rather than
  a regression. Established twice over: the board runs the default module tree
  with no MoonLive module at all, and a firmware built from a clean main crashes
  identically. bootReason alternates PANIC and WDT because the panic handler
  sometimes dies before finishing its dump. The fault site is FreeRTOS's RISC-V
  coprocessor save on interrupt entry, and the prior art at Plan-20260718 is a
  different cause with the same panic name, worth re-reading first.
- Plan-20260817: steps 5 and 10 are done, both by the work in Plan-20260818, and
  the arena ceiling has its number. Only `get(x, y)` is open now. Step 10's own
  reasoning is recorded as wrong: it said the editing loop should come last,
  against a finished language, but the shape an editor needs does not change with
  the language, and it paid for itself immediately.
- Plan-20260818 records what the plan did not predict: the seam needed nothing
  built, the editor already existed, and two language decisions were corrected
  during the work.

Tests
- A uint16_t member starts at the value it was initialized to, observed through a
  COMPARISON rather than a colour channel, which is why the bug survived until
  now. Control-checked: restoring the cast fails it.

Bench: all four boards flashed and re-seeded with the 17 role-named scripts. The
P4 boots and holds its scripts for the first time on this branch.

+++ src/core/moonlive/MoonLiveCompiler.cpp
@@ -616,7 +616,10 @@ struct Parser {
         if (align > 1 && (at % align) != 0) at = uint16_t(at + (align - at % align));
         const uint16_t need = uint16_t(ctrlWidth(type));
         if (at + need > kCtrlBytes) { fail("the class declares more member data than the arena holds"); return; }
-        members[memberCount] = {name, 0, 255, static_cast<uint8_t>(def),
+        // def is uint16_t on the record precisely so a wide member's initializer survives; casting
+        // it to a byte here truncated `uint16_t phase = 1000;` to 232. Invisible to a test that
+        // observes through setRGB, because the error is always a multiple of 256.
+        members[memberCount] = {name, 0, 255, static_cast<uint16_t>(def),
                                 static_cast<uint8_t>(nameLen), type,
                                 static_cast<uint8_t>(at), 1};
         memberBytes = static_cast<uint8_t>(at + need);
